fetcher/requirements.txt
amqp==5.1.1
async-timeout==4.0.2
asyncio==3.4.3
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
flower==1.0.0
gevent==21.12.0
greenlet==1.1.2
grpcio==1.46.3
grpcio-tools==1.46.3
humanize==4.1.0
kombu==5.2.4
motor==3.0.0
packaging==21.3
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
pymongo==4.1.1
pyparsing==3.0.9
pytz==2022.1
redis==4.3.3
six==1.16.0
tornado==6.1
vine==5.0.0
wcwidth==0.2.5
websockets==10.3
wrapt==1.14.1
zope.event==4.5.0
zope.interface==5.4.0
